+++ Packages/com.nickmaltbie.openkcc.netcode/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## In Progress
 
+* Added controls for loading and configuring ssl connection
+    * Also added field for supplying a private key for the server hosting.
 * Change port configuration for web socket to use secure connection over https.
 * Added example using unity Netcode and sub package com.nickmaltbie.openkcc.netcode
     * Added example NetworkKCC for a networked kinematic character controller.